MainPage Добавил передачу данных на экран оформления займа

+++ app/src/main/java/com/pogreb/shift_pogrebiczkij_2025/feature/main_page/domain/entity/LoanConditions.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 data class LoanConditions(
     val percent: Double,
     val period: Int,
-    val maxAmount: Int,
+    val maxAmount: Long,
 )
